renamed table extracted efa model classes
