+++ app/src/main/java/de/tum/in/tumcampusapp/managers/TransportManager.java
@@ -20,17 +20,17 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.GregorianCalendar;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.TimeUnit;
 
 import de.tum.in.tumcampusapp.auxiliary.Const;
 import de.tum.in.tumcampusapp.auxiliary.NetUtils;
 import de.tum.in.tumcampusapp.auxiliary.Utils;
 import de.tum.in.tumcampusapp.cards.MVVCard;
 import de.tum.in.tumcampusapp.cards.generic.Card;
-import de.tum.in.tumcampusapp.widgets.MVVWidget;
+import de.tum.in.tumcampusapp.models.efa.Departure;
+import de.tum.in.tumcampusapp.models.efa.StationResult;
+import de.tum.in.tumcampusapp.models.efa.WidgetDepartures;
 
 /**
  * Transport Manager, handles querying data from mvv and card creation
@@ -125,7 +125,7 @@ public class TransportManager extends AbstractManager implements Card.ProvidesCa
         db.execSQL("CREATE TABLE IF NOT EXISTS transport_favorites (" +
                 "id INTEGER PRIMARY KEY AUTOINCREMENT, symbol VARCHAR)");
 
-        db.execSQL("CREATE TABLE IF NOT EXISTS transport_widgets (" +
+        db.execSQL("CREATE TABLE IF NOT EXISTS widgets_transport (" +
                 "id INTEGER PRIMARY KEY, station VARCHAR, station_id VARCHAR, location BOOLEAN, reload BOOLEAN)");
 
         if(TransportManager.widgetDeparturesList == null) {
@@ -168,7 +168,7 @@ public class TransportManager extends AbstractManager implements Card.ProvidesCa
         values.put("station_id", widgetDepartures.getStationId());
         values.put("location", widgetDepartures.useLocation());
         values.put("reload", widgetDepartures.autoReload());
-        db.replace("transport_widgets", null, values);
+        db.replace("widgets_transport", null, values);
         TransportManager.widgetDeparturesList.put(appWidgetId, widgetDepartures);
     }
 
@@ -178,7 +178,7 @@ public class TransportManager extends AbstractManager implements Card.ProvidesCa
      * @param widget_id The id of the widget
      */
     public void deleteWidget(int widget_id) {
-        db.delete("transport_widgets", "id = ?", new String[]{String.valueOf(widget_id)});
+        db.delete("widgets_transport", "id = ?", new String[]{String.valueOf(widget_id)});
         TransportManager.widgetDeparturesList.remove(widget_id);
     }
 
@@ -196,7 +196,7 @@ public class TransportManager extends AbstractManager implements Card.ProvidesCa
         if(TransportManager.widgetDeparturesList.indexOfKey(widget_id) >= 0){
             return TransportManager.widgetDeparturesList.get(widget_id);
         }
-        Cursor c = db.rawQuery("SELECT * FROM transport_widgets WHERE id = ?", new String[]{String.valueOf(widget_id)});
+        Cursor c = db.rawQuery("SELECT * FROM widgets_transport WHERE id = ?", new String[]{String.valueOf(widget_id)});
         WidgetDepartures widgetDepartures = new WidgetDepartures();
         if (c.getCount() >= 1) {
             c.moveToFirst();
@@ -359,167 +359,5 @@ public class TransportManager extends AbstractManager implements Card.ProvidesCa
         card.setDepartures(cur);
         card.apply();
 
-    }
-
-    public static class Departure {
-        final public String servingLine;
-        final public String direction;
-        final public String symbol;
-        final public int countDown;
-        final public long departureTime;
-
-        public Departure(String servingLine, String direction, String symbol, int countDown, long departureTime) {
-            this.servingLine = servingLine;
-            this.direction = direction;
-            this.symbol = symbol;
-            this.countDown = countDown;
-            this.departureTime = departureTime;
-        }
-
-        /**
-         * Calculates the countDown with the real departure time and the current time
-         *
-         * @return The calculated countDown in minutes
-         */
-        public long getCalculatedCountDown(){
-            return TimeUnit.MINUTES.convert(departureTime - System.currentTimeMillis(), TimeUnit.MILLISECONDS);
-        }
-    }
-
-    public static class StationResult {
-        final String station;
-        final String id;
-        final int quality;
-
-        public StationResult(String station, String id, int quality) {
-            this.station = station;
-            this.id = id;
-            this.quality = quality;
-        }
-    }
-
-    public static class WidgetDepartures {
-
-        private String station;
-        private String station_id;
-        private boolean use_location;
-        private boolean auto_reload;
-        private List<Departure> departures;
-        private long last_load;
-        private boolean is_offline = false;
-
-        /**
-         * Create a new WidgetDepartures. It contains the widget settings and can load the according departure list
-         */
-        WidgetDepartures() {
-            this.station = "";
-            this.station_id = "";
-            this.auto_reload = false;
-            this.use_location = false;
-            this.departures = new ArrayList<>();
-        }
-
-        /**
-         * The station_id which is set for this widget
-         * @return The station name
-         */
-        public String getStationId() {
-            return this.station_id;
-        }
-
-        /**
-         * Sets a station_id for this widget
-         * @param station_id The station name
-         */
-        public void setStationId(String station_id) {
-            if(!this.station_id.equals(station_id)){
-                this.departures.clear();
-            }
-            this.station_id = station_id;
-        }
-
-        /**
-         * The station which is set for this widget
-         * @return The station name
-         */
-        public String getStation() {
-            if (this.use_location) {
-                // TODO implement nearest station (replace the station_id string with the calculated station)
-                this.station = "use location";
-            }
-            return this.station;
-        }
-
-        /**
-         * Sets a station title for this widget
-         * @param station The station name
-         */
-        public void setStation(String station) { this.station = station;}
-
-        /**
-         * Whether this widgets station is determined by the current location
-         * @return True if location is used
-         */
-        boolean useLocation() { return use_location; }
-
-        /**
-         * Whether this widgets station should determined by the current location
-         * @param use_location True is location has to be used
-         */
-        void setUseLocation(boolean use_location) { this.use_location = use_location; }
-
-        /**
-         * True if widget should update automatically, otherwise a button-press is required
-         * @return Whether auto_reload is enabled
-         */
-        public boolean autoReload() {
-            return this.auto_reload;
-        }
-
-        /**
-         * True if widget should update automatically, otherwise a button-press is required
-         * @param auto_reload Whether auto_reload should enabled
-         */
-        public void setAutoReload(boolean auto_reload) { this.auto_reload = auto_reload; }
-
-        /**
-         * Are the departure information older than two minutes (because of any connection problems)
-         *
-         * @return True if only offline data available
-         */
-        public boolean isOffline(){
-            return this.is_offline;
-        }
-
-        /**
-         * Get the list of departures for this widget, download them if they are not cached
-         *
-         * @return The list of departures
-         */
-        public List<Departure> getDepartures(Context context, boolean force_server_load) {
-            if(this.departures == null){
-                this.departures = new ArrayList<>();
-            }
-            // download only id there is no data or the last loading is more than X min ago
-            if (this.departures.size() == 0 || force_server_load || (this.autoReload() && System.currentTimeMillis() - this.last_load > MVVWidget.DOWNLOAD_DELAY)) {
-                List<Departure> departures = TransportManager.getDeparturesFromExternal(context, this.getStationId());
-                if(departures.size() == 0){
-                    this.is_offline = true;
-                } else {
-                    this.departures = departures;
-                    this.last_load = System.currentTimeMillis();
-                    this.is_offline = false;
-                }
-            }
-
-            // remove Departures which have a negative countdown
-            for (Iterator<Departure> iterator = this.departures.iterator(); iterator.hasNext();) {
-                Departure departure = iterator.next();
-                if (departure.getCalculatedCountDown() < 0) {
-                    iterator.remove();
-                }
-            }
-            return this.departures;
-        }
     }
 }
